Replace StartSubscriptionCommand with ProviderBillingService and moved to Commercial
